rag_system/requirements.txt
sentence-transformers==2.2.2
huggingface-hub==0.16.4
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.24.3
torch==2.0.1
transformers==4.30.2 